final demo update for FFT
